add string run

+++ src/lib/ser.h
@@ -5,10 +5,10 @@
 
 typedef struct _ser_buffer {
     MEM_HEADER(_ser_buffer);
-    size_t size;
+    uint32_t size;
     uint8_t data[];
 } ser_buffer;
 
-ser_buffer *ser_buffer_init(size_t size);
+ser_buffer *ser_buffer_init(uint32_t size);
 
 void ser_buffer_free(ser_buffer *buffer);
